features, changes and fixes
feat(user): UserModelRegister and validations added to the users aplication.
This model was added to separate the registration logic and data insertion into the database data collection.

feat(categories): The category model was created for use in relation to products.

feat(categories): A CRUD of categories was added for management