fixes
fix(database): Correction of the name of the user collection.

fix(auth): A function was deleted in the user model that was repeated for different fields.
In made a function to validate that the password and confirm_password fields have the special characteristics

feat: Added "cherrypy" for the use of WSGI servers.

fix: Dockerfile changes.

docs: added readme